Updated Javadoc in multiply method (feature branch)

+++ src/main/java/labxx/sict/must/edu/mn/Multiplication.java
@@ -1,8 +1,15 @@
 package labxx.sict.must.edu.mn;
 
 public class Multiplication {
-    public double multiply(double a, double b) {
-        return a * b;
-    }
+   /**
+ * Multiplies two integers and returns the product.
+ * @param a first integer
+ * @param b second integer
+ * @return product of a and b
+ */
+public int multiply(int a, int b) {
+    return a * b;
+}
+
 }
